+++ platformio/package/commands/install.py
@@ -23,7 +23,9 @@ from platformio.package.exception import UnknownPackageError
 from platformio.package.manager.library import LibraryPackageManager
 from platformio.package.manager.platform import PlatformPackageManager
 from platformio.package.manager.tool import ToolPackageManager
-from platformio.package.meta import PackageSpec
+from platformio.package.meta import PackageCompatibility, PackageSpec
+from platformio.platform.exception import UnknownPlatform
+from platformio.platform.factory import PlatformFactory
 from platformio.project.config import ProjectConfig
 from platformio.project.savedeps import pkg_to_save_spec, save_project_dependencies
 from platformio.test.result import TestSuite
@@ -202,8 +204,24 @@ def _install_project_env_libraries(project_env, options):
     _uninstall_project_unused_libdeps(project_env, options)
     already_up_to_date = not options.get("force")
     config = ProjectConfig.get_instance()
+
+    compatibility_qualifiers = {}
+    if config.get(f"env:{project_env}", "platform"):
+        try:
+            p = PlatformFactory.new(config.get(f"env:{project_env}", "platform"))
+            compatibility_qualifiers["platforms"] = [p.name]
+        except UnknownPlatform:
+            pass
+        if config.get(f"env:{project_env}", "framework"):
+            compatibility_qualifiers["frameworks"] = config.get(
+                f"env:{project_env}", "framework"
+            )
+
     env_lm = LibraryPackageManager(
-        os.path.join(config.get("platformio", "libdeps_dir"), project_env)
+        os.path.join(config.get("platformio", "libdeps_dir"), project_env),
+        compatibility=PackageCompatibility(**compatibility_qualifiers)
+        if compatibility_qualifiers
+        else None,
     )
     private_lm = LibraryPackageManager(
         os.path.join(config.get("platformio", "lib_dir"))